requirements.txt
Flask
google-cloud-datastore
google-auth
requests
google-cloud-secret-manager
google-cloud-storage
Pillow
pytest
